SecureTalk/requirements.txt
Flask==2.3.3
Flask-Limiter==3.0.1
pycryptodome==3.18.0